feat: added mint code and refactoring send token code
